refactor(tui): keep completion changes off the code-health gate
Rebasing onto main needed three fixes, and the shape of them was driven
by CodeScene, which fails the build if any touched file's health drops.

`.info mcp` no longer exists: the direct MCP path went away with the
McpManager removal, so only `.info env` keeps a usage hint. `Tui::init`
also lost its hook-manager arguments, which the new test still passed.

The rest is arranging the change so it does not degrade three files:

- The dedup for first-word completions moved out of
  `compute_completions`, which was already past the complexity
  threshold, into `completion::command_name_completions`. Putting it in
  its own module rather than in `input.rs` matters -- as a free function
  there it pushed the module to 76 functions against a threshold of 75.
- `command_complete` grew two more match arms of fixed subcommand
  lists. Those arms are data, so all seven of them became a
  `FIXED_SUBCOMMANDS` table and one fallback arm, which leaves the match
  holding only the cases that compute something.
- `tests.rs` is 6889 lines and 231 functions, both far over threshold,
  so the new test went into `tests/command_completion.rs` instead.

`cs delta origin/main` now reports input.rs level at 3.14 with fewer
lines, and no other file flagged.
