+++ crates/harnx-tui/src/lib.rs
@@ -16,6 +16,8 @@ pub mod terminal;
 pub mod test_utils;
 pub mod types;
 
+mod completion;
+
 #[cfg(test)]
 mod tests;
